requirements.txt
Flask
Flask-MySQL
Flask-WTF
mysql-connector-python
